Remove stream id module for now
