+++ crates/metrics/src/stats.rs
@@ -12,6 +12,8 @@ use std::{
     iter::FromIterator,
 };
 
+use regex::Regex;
+
 pub trait DescriptiveStats {
     fn max(&self) -> f64;
     fn min(&self) -> f64;
@@ -547,6 +549,27 @@ impl StatsByMetric<StatsRecord> {
                 stats_by_metric_name
             },
         ))
+    }
+
+    pub fn group_by_regex<I: IntoIterator<Item = Metric>>(
+        re: &Regex,
+        metrics: I,
+    ) -> StatsByMetric<OnlineStats> {
+        StatsByMetric(metrics.into_iter().fold(HashMap::new(), |mut map, metric| {
+            let metric_name = metric.name.clone();
+            let stream_id = metric.stream_id.clone();
+            stream_id
+                .and_then(|stream_id| {
+                    re.captures_iter(stream_id.as_str()).next().map(|captured| {
+                        let key = GroupingKey::new(captured[1].to_string(), metric_name);
+                        let entry = map.entry(key);
+                        let stats: &mut OnlineStats = entry.or_insert_with(OnlineStats::empty);
+                        stats.add(metric.value)
+                    })
+                })
+                .unwrap_or_else(|| {});
+            map
+        }))
     }
 }
